Set up the App.js so either AuthNav.js or AppNav.js shows to user
